Update patch set 6
Patch Set 6:

Some additional notes: 

I was working on an asm implementation for poly1305 on ppc64le based on the s390x implementation and it will be ready to be submitted soon.

In looking at few profiles it looks like one of the performance problems is due to the loading and storing of arguments and return values for add128. Should that function be inlined?

We are looking at making bits.Add64 an intrinsic which should also help.

Patch-set: 6