x/crypto/sha3: fix bug in cSHAKE Clone()
Clone() copied the Keccak state directly, which is not supported, since the "buf"
slice in the Keccak state must point to the "storage" array, and that doesn't happen
if you copy the state directly. It was fixed by embedding a pointer to the Keccak state
instead of the state itself.

Change-Id: I7d392963ec65d784a360f6c12a7935a9a9a788b5