Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> (1 comment)
> 
> Please rebase on top of the cSHAKE change that was just merged.

Done.

While doing this I found a bug in the cSHAKE implementation that I fixed here; I was not sure if I should open a separate issue/CL. Let me know if that's preferable. (Basically it copied the Keccak state directly, which is not supported, since the "buf" slice in the Keccak state must point to the "storage" array, and that doesn't happen if you copy the state directly. I fixed it by embedding a pointer to the state instead of the state itself. I also added a test that triggered this bug.)

I also added a KMAC example.

Patch-set: 3
Reviewer: Gerrit User 11965 <11965@62eb7196-b449-3ce5-99f1-c037f21e1705>