Update patch set 1
Patch Set 1:

Mike, I've tried using the MAC implementation for this but running into a snag when the message is large but has a leftover which is not a multiple of 16. In that case it saves intermediate values in state which are then used on the next call to update. Trying to work through that now, since the amd64 implementation did not use vectors like ours. Seems like there are other values from the call to update on the first part of the message that need to be available when processing the tail.

Patch-set: 1
Reviewer: Gerrit User 6320 <6320@62eb7196-b449-3ce5-99f1-c037f21e1705>