Update patch set 7
Patch Set 7:

(4 comments)

The last patch-set moved the server side implementation to _test.go so we do not longer provide a working server implementation. Currently it panics, if the server side of this kex is used. I personally think, that panic is the right call here (alternatively an error could be returned). That being said, I thought about the right place to add this to the documentation, when it occured to me, that there is currently nothing about the supported kex in the documentation anyways so everyone interested in providing his own set of kex needs to lookup the code anyways. So I added a comment where the kexAlgo constants are defined.

All the other comments I have resolved with one exception: I do not understand what "f" is, where I am supposed to add additional checks for it being > 1 and < (p - 1) / 2.

Patch-set: 7