Update patch set 8
Patch Set 8:

> Patch Set 7:
> 
> (4 comments)
> 
> The last patch-set moved the server side implementation to _test.go so we do not longer provide a working server implementation. Currently it panics, if the server side of this kex is used. I personally think, that panic is the right call here (alternatively an error could be returned). That being said, I thought about the right place to add this to the documentation, when it occured to me, that there is currently nothing about the supported kex in the documentation anyways so everyone interested in providing his own set of kex needs to lookup the code anyways. So I added a comment where the kexAlgo constants are defined.
> 

panic is almost certainly the wrong solution: it will take the server down, and puts this under client control. I also couldn't see where the server code would actually be exercised. 

Can you move the code back to sit next to the client code? This helps with understanding what is going on.

To enforce that it's not used in production, check for the forbidden kex at the start of NewServerConn (or whereever else we check ServerConfig.Config for saneness.)

> All the other comments I have resolved with one exception: I do not understand what "f" is, where I am supposed to add additional checks for it being > 1 and < (p - 1) / 2.

Patch-set: 8
Reviewer: Gerrit User 5893 <5893@62eb7196-b449-3ce5-99f1-c037f21e1705>