Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> still needs a unittest. Did you check coverage?

I don't understand what exactly you mean with "still needs a unittest". The kex is tested by the existing unit test in kex_test.go. The coverage (for /x/crypto/ssh) without this patch is 77.4% and with the patch is 77.3%. The slight decrease is because the existing unittest does not cover all the possible code paths. Because of the pure nature of this kex (it performs a negotiation where different outcomes are possible) there are quite some angles to cover.

In order to cover these angles, we would need different behaving server implementations, because the client half (and I assume, we mainly care about the client half, because we prevent the usage of the server half in the implementation) can essentially only be tested with a respective server half (which provides a certain behavior).

That being said, in Patchset 8 I provided an implementation, which would let me swap the server half during tests and therefore cover the different angles, but you requested to revert this change and have the server next to the client the same way it is done for the other kex.

So now, I really do not understand, what exactly you want me to provide in order to get approval for this PR.

Patch-set: 9