Update patch set 4
Patch Set 4:

Brad, this patch originally had concerns due to its licensing related to OpenSSL and we were waiting on some code to be put into the cryptogams repository. Now that it is there, it is our understanding that it is acceptable. Is this now OK?

Patch-set: 4
CC: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>
Reviewer: Gerrit User 6320 <6320@62eb7196-b449-3ce5-99f1-c037f21e1705>