Update patch set 12
Patch Set 12:

> Patch Set 11:
> 
> you can try editing the message inside Gerrit directly, using the "edit commit message" button.

This is very unfortunate, that the sync between github and gerrit does not cover the commit message.

Finally I got this to work as well, I first had to fix a problem with the e-mail address used in the commit in order to be allowed to change the git commit message in gerrit.

Patch-set: 12