Update patch set 1
Patch Set 1:

Hi, and thanks for contributing. Adding a new package would require first a proposal on the issue tracker, but I can tell you in advance that it's unlikely we'll expose these algorithms in x/crypto. The point of x/crypto is not to provide all possible algorithms, but a curated, useful, opinionated and safe subset. KDF1/2 would be redundant with HKDF, and they aren't widespread or complex enough to justify the redundancy. They would be surely helpful to the community on GitHub (or any other host) though, not everything has to be in x/crypto!

Patch-set: 1
Reviewer: Gerrit User 11715 <11715@62eb7196-b449-3ce5-99f1-c037f21e1705>