scrypt: explain how to generate a salt
The example explains that you are not supposed to use the provided
salt, but the first recourse of people pressed for time is going
to be to do the easiest thing possible, copy the salt from the
demonstration.

Show users how to generate their own salts, instead of copying the one
from the example.

Change-Id: I432173a7a3c2962735bb0ce5549c33b49a1da400