Update patch set 5
Patch Set 5:

@Han-Wen respectfully, I think you should reconsider your stance on this. I agree that a ControlMaster implementation itself shouldn't be part of this library, however exposing Transport seems like a very reasonable compromise here. If not, this isn't even something that can be solved via external packages as there is no way for an external package to satisfy the `*ssh.Client` interface that other external packages may expect (ex: sftp) because it's a struct within internal fields not an interface. 

Since this request was spawned off a discussion on ControlMaster, I also wanted to give a real-world example of ControlMaster usage where mixing OpenSSH with go could be reasonable:

The use of ControlMaster to enforce MFA but still provide a low friction developer experience is very common in the enterprise world. A typical setup may look like this (configured via an OpenSSH ssh_config file):

> workstation -> bastion (MFA) -> production host

ControlMaster is enabled for the bastion host so that the user establishes an active connection once (completing MFA) to the bastion host and then re-uses it for every ongoing request to production hosts allowing them too seamlessly "ssh foo.prod.enterprise.com" without re-prompting for MFA.

Now imagine a go CLI has been written that takes a hostname and runs a command on the host. Using this package as-is, that would mean MFA must be entered on every invocation of the CLI. Alternatively, and entire server/RPC mechanism could be written and forked, but that seems like unnecessary work (and potentially is re-implemented for every CLI), and will still require an extra MFA on first connection. It would be great to be able to specifically use an already established ControlMaster via this Transport method.

Patch-set: 5
CC: Gerrit User 31156 <31156@62eb7196-b449-3ce5-99f1-c037f21e1705>