Update patch set 8
Patch Set 8:

(2 comments)

> Patch Set 7: Code-Review-1
> 
> (1 comment)

Thanks for the review! I added the non-valid signature tests and added comments.

Regarding the implementation of u2f/fido2 functionality, this would only be necessary if we wanted to perform client actions (or test the flow end-to-end).

Our priority is getting server-side support for SSH-SK authentication into x/crypto/ssh, as we see this as the most likely used configuration (go SSH server, openssh client). 

Adding client support to x/crypto/ssh would require talking to a USB device, or at least providing interfaces which can be piped through. That's something I'd like to do in the future, but won't be able to do at the moment.

Patch-set: 8