Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> (2 comments)
> 
> > Patch Set 7: Code-Review-1
> > 
> > (1 comment)
> 
> Thanks for the review! I added the non-valid signature tests and added comments.
> 
> Regarding the implementation of u2f/fido2 functionality, this would only be necessary if we wanted to perform client actions (or test the flow end-to-end).

This is about testing the flow end-to-end. If this is reasonably simple this may be worth it; it would also be instructive for people that want to understand how U2F for SSH works.

> Our priority is getting server-side support for SSH-SK authentication into x/crypto/ssh, as we see this as the most likely used configuration (go SSH server, openssh client). 
>
> Adding client support to x/crypto/ssh would require talking to a USB device, or at least providing interfaces which can be piped through. That's something I'd like to do in the future, but won't be able to do at the moment.

Yes, of course - I agree with this.

Patch-set: 8
Reviewer: Gerrit User 5893 <5893@62eb7196-b449-3ce5-99f1-c037f21e1705>