Update patch set 7
Patch Set 7:

(1 comment)

> Brainstormed with Filippo, Jay, and Rebecca and we've come up with the solution of running `go mod download -m -json` in a test main, and using the Dir in the JSON given back to read the json tests we need.

I think that's a great solution, nice work! I think it addresses a lot of the problems with trying to actually require that module in go.mod that Heschi made me aware of (e.g., go mod tidy removing it), and it's going to be fast/efficient to test often because of the module cache.

I just left one minor simplification driveby comment.

Patch-set: 7
CC: Gerrit User 17090 <17090@62eb7196-b449-3ce5-99f1-c037f21e1705>