Update patch set 8
Patch Set 8:

> Patch Set 7:
> 
> (1 comment)
> 
> > Brainstormed with Filippo, Jay, and Rebecca and we've come up with the solution of running `go mod download -m -json` in a test main, and using the Dir in the JSON given back to read the json tests we need.
> 
> I think that's a great solution, nice work! I think it addresses a lot of the problems with trying to actually require that module in go.mod that Heschi made me aware of (e.g., go mod tidy removing it), and it's going to be fast/efficient to test often because of the module cache.

It's better in that respect, but still leaves a fair bit to be desired. It won't work in an environment without network access, or where the go command isn't available. WASM is one obvious example. I don't know if these tests are relevant there.

Have you considered bundling the test data into an actual Go package somehow and depending on it as a real module?

Patch-set: 8