Update patch set 3
Patch Set 3:

Thanks for the ping. I realized that there is a bit of a footgun here: if SetCounter is called with a counter lower than the current one it's a guaranteed keystream reuse.

How about adding a check against it and documenting it?

if counter < s.counter {
    panic("chacha20: SetCounter attempted to rollback counter")
}

(I think the check should be <, not <=, right? Because the counter is that of the next iteration, and if called with the current counter SetCounter will just discard the buffer, which is what Advance used to do, so it's also going to be useful in chacha20poly1305.)

At that point, should we call it AdvanceCounter? Or even Advance?

Patch-set: 3
Reviewer: Gerrit User 11715 <11715@62eb7196-b449-3ce5-99f1-c037f21e1705>