Update patch set 5
Patch Set 5:

> Patch Set 4: TryBot-Result-1
> 
> 2 of 10 TryBots failed:
> Failed on linux-386: https://storage.googleapis.com/go-build-log/f770366f/linux-386_e2187f46.log
> Failed on windows-386-2008: https://storage.googleapis.com/go-build-log/f770366f/windows-386-2008_cc485d72.log

Both the P-256 and the P-384 test keys have an extra leading zero byte which is causing this crash in the generic implementations, because it expects the input not to be longer than the scalar size in bytes.

Any idea why that zero byte is there? These are the keys:
0081c1e3c47e82865a8979dd7d84ebc24e014f5f8c78b3c2dbf7707b4279775403
00e0d21c06281ffd12af9ed85793b298bc4075b9c530faa6baac1c3171e1c2a4d1c720155eb57611610fff1211b6251f31

Patch-set: 5
Reviewer: Gerrit User 11715 <11715@62eb7196-b449-3ce5-99f1-c037f21e1705>