Update patch set 5
Patch Set 5: Run-TryBot+1 Code-Review+2

> Patch Set 4:
> 
> (1 comment)
> 
> After some thinking and digging around I'm convinced that the lack of bounds check elimination is actually the intended behaviour, because when you do foo[:x], this is permitted so long as x <= cap(foo) (not len(foo)). Checking the cap gives the compiler no information about its len, so all subsequent accesses to foo (even foo[0]) are still not guaranteed to be safe.
> 
> I realised this when I read this: https://go101.org/article/bounds-check-elimination.html
> This demonstrates some panics: https://play.golang.org/p/mUaFA8jQVPG

Oooooh that makes sense!

Then the thing that would work is to assign the result back to the slice, or to a new temporary one and use the latter. If SSA is doing its job perfectly, that should not have any overhead, but I'm willing to believe it doesn't.

I'm curious to know if you want to try it out, but otherwise I am happy with this CL as is. Thanks again!

Patch-set: 5
Reviewer: Gerrit User 11715 <11715@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2
Label: Run-TryBot=+1