Update patch set 10
Patch Set 10:

(3 comments)

Thank you for the detailed comments!

I saw you had mentioned experimenting with code generation for this. If it didn't work out, could you share why and what's there to learn? (I'm ok with landing this as assembly, since it improves on the current assembly and simplifies the rest of the code.)

There are some units that might be worth testing individually, and which might make it possible to test for carry edge cases: MULTIPLY and REDUCE for example. I don't want to push you past the freeze, so I'm ok with a follow-up CL if you think it's indeed feasible and useful.

Patch-set: 10
Reviewer: Gerrit User 11715 <11715@62eb7196-b449-3ce5-99f1-c037f21e1705>