Update patch set 11
Patch Set 11: Run-TryBot+1

(3 comments)

> Thank you for the detailed comments!

Thanks (to you and Lynn) for reviewing it! I know CLs like this one can be a bit of a slog.

> I saw you had mentioned experimenting with code generation for this. If it didn't work out, could you share why and what's there to learn? (I'm ok with landing this as assembly, since it improves on the current assembly and simplifies the rest of the code.)

It was going OK, I paused working on it when I went on paternity leave and haven't resumed yet, so it isn't finished. Poly1305 is an interesting case because it doesn't have a lot of repetition outside of multiply/reduce so codegen doesn't make a massive difference to readability. Automatic register allocation and deallocation does make tracking register usage much easier however and I really liked the impact that had.

Overall I'm happy with the experience so far and I'd like to start converting more assembly to be generated, though I think I might start with chacha20 rather than poly1305 since it will need fewer instructions and will benefit more due to the amount of repetition. Plus I'd like to push chacha20 performance a bit more since it is dominating the AEAD performance now and using a code generator will make it easier to play with the number of blocks calculated in parallel.

One unsolved problem is how to package up the code generator as I add features to it. I'd like to keep some stuff in its own package separate from the target assembly.

> There are some units that might be worth testing individually, and which might make it possible to test for carry edge cases: MULTIPLY and REDUCE for example. I don't want to push you past the freeze, so I'm ok with a follow-up CL if you think it's indeed feasible and useful.

Yeah, I was pondering this too. MULTIPLY and REDUCE would be fairly easy to unit test, but I don't think they would benefit much from it since they are actually quite simple and tested well by the black box testing. Other parts of the code are much more susceptible to problems with edge cases that do not arise in randomized testing (mostly truncating very-unlikely-to-be-set high bits in limbs). This is why I added the initial state tests for example: the state never exceeds 2**130 - 1 in the existing test cases so initial state loading wasn't fully tested.

Patch-set: 11
Reviewer: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Run-TryBot=+1