Update patch set 5
Patch Set 5:

(I was browsing through open CLs and stumbled on this one. This is just my own opinion; i don't speak for the Go team.)

The linked RFC simply says that Kerberos 5 can express a PBKDF2 iteration count between 1 and 2**32. It doesn't say that implementations have to support this whole range, and indeed it says the opposite - that implementations might want to support a smaller range. 

   For environments where slower hardware is the norm, implementations
   [...] may wish to limit the number of iterations to prevent a spoofed
   response supplied by an attacker from consuming lots of client-side
   CPU time.

On my laptop, 2^32 iterations of PBKDF2-HMAC-SHA1 would take an hour or two, which is probably overkill for an interactive authentication protocol like Kerberos. I don't think iteration counts above 2^31 are likely to be seen in practice. (But if you have evidence to the contrary please share.)

Maybe pbkdf2.Key should have accepted an int64 iteration count in the first place, but we can't change it now, and it doesn't seem worth it to me to add new API just to cover a corner case that only exists on 32-bit platforms and which probably isn't needed in practice.

Patch-set: 5
CC: Gerrit User 6255 <6255@62eb7196-b449-3ce5-99f1-c037f21e1705>