x/crypto/acme: make Client.Register behavior consistent
This change makes the behavior of Client.Register consistent for both
pre-8555 and 8555 implementations consistent when the account already
exists. The 8555 implementation is altered to match the pre-8555
behavior, if the account exists the Account object is returned without
an error. Most consumers I surveyed treat err == nil and err ==
ErrAccountAlreadyExists the same, an example of which was in autocert,
and as such simply transparently returning the Account without this
error _shouldn't_ cause any issues for existing users. The error that
previously flagged this, ErrAccountAlreadyExists, is removed.

The existing 8555 Register test is updated to check that both the new
account and existing account behavior matches, and a previous test for
the special already exists behavior is removed.

Change-Id: I2b7944baf25b70ccb7d12c5c8e1fc6f528d09f3d