Update patch set 2
Patch Set 2:

I've added the requested tests I hope these work.

As a side note it seems the original code checks if the supplied cipher list is actually supported and removes the ciphers configured but not supported silently. My implementation for MACs and Kex' currently throws an error. I feel that there is a bit of a dissonance there and it would be good if the handling would be the same. I see three options:

1. Throw errors for ciphers too
2. Change MAC and Kex checks to silent drop unsupported items
3. Change MAC and Kex to filter and add a log warning to all three.

Patch-set: 2