Update patch set 5
Patch Set 5:

David, out of curiosity, do you have a desire to be able to swap out the transport on the server interface as well? Not trying to expand the scope of an already contentious change just yet, but given the focus of your employer I figured I would ask.

Patch-set: 5