x/crypto/cryptobyte: clarify ReadOptionalASN1Boolean doc
Brings it in line with the other the ReadOptionalASN1* method docs.
These methods are in general a little confusing because they only
work for OPTIONAL/DEFAULT fields that use context-specific tags.
Probably the documentation could be further improved to make this
clearer, and suggest what to do for OPTIONAL/DEFAULT fields that
aren't using context-specific tags.

Change-Id: I0b2cde6ed2337946f50bb169618807e0945b81bd