curve25519: add GenerateKeypair helper function
One of the most common uses of the curve25519 package is simply
generating a keypair, storing the private key, and shipping the public
key off to some service. People seem to do this in different ways, and
sometimes forget to clamp the private keys, which generally isn't a
problem, but can make resultant behaviors surprising to administrators
or make less robust code fail equality tests. It's also not uncommon to
see people forgetting to test the return value of rand.Read. Since this
is such a common operation, it makes sense to provide a trivial helper
function that generates a clamped scalar and the corresponding public
key, all in a single easy to use function, with no surprises. This
commit also adds a test case that makes sure the output works for basic
diffie-hellman.

Change-Id: Ifa69440e1651ddfa952fe32c3cce13567b4059cc