curve25519: don't allow basePoint bytes to be mangled
Rather than exposing a security-critical constant to the outside of this
package, this commit always uses a private symbol for the actual bytes.
The previous code added extra overhead of doing a constant time byte
comparison to the original base point. Rather than having the carry two
constants around -- the one used for calculations and the one used for
checking that the first is okay -- we can have a single private source
of truth, by making the Basepoint exported symbol be nothing more than a
placeholder token that has no value for the calculation.

Change-Id: I1f5763fa718b8c469390a47904105a62467adafb