crypto/salsa20/salsa: implement XORKeyStream12 and XORKeyStream8
Implementing the 12 and 8 in both go and assembly.

For assembly, I did minimal changes here by using the amd64 BX register for the round count, it's limited to uint16, but that's enough for 20, 12, and 8.  ;)

All the best!

Change-Id: I33bba926e860b7505e0ee055c5fcf0ed57218d9d
GitHub-Last-Rev: 8412a7f757e7abca5baaabd167d9a7ff1518413d
GitHub-Pull-Request: golang/crypto#174